+++ dpws/src/main/java/org/somda/sdc/dpws/soap/wsdiscovery/WsDiscoveryClient.java
@@ -10,18 +10,34 @@ import org.somda.sdc.dpws.soap.wsaddressing.model.EndpointReferenceType;
 import org.somda.sdc.dpws.soap.wsdiscovery.model.ProbeMatchesType;
 import org.somda.sdc.dpws.soap.wsdiscovery.model.ResolveMatchesType;
 
+import javax.annotation.Nullable;
 import javax.xml.namespace.QName;
 import java.util.Collection;
 import java.util.List;
 
 /**
- * Ws-Discovery Client interface.
+ * WS-Discovery Client interface.
  *
  * @see <a href="http://docs.oasis-open.org/ws-dd/discovery/1.1/os/wsdd-discovery-1.1-spec-os.html#_Toc234231815"
  * >Conceptual Message Content</a>
  */
 public interface WsDiscoveryClient extends Interceptor {
 
+    /**
+     * Sends a directed probe to a device.
+     *
+     * @param rrClient to send request on
+     * @param types    to probe for
+     * @param scopes   to probe for
+     * @param matchBy the scopes matching rule, or null to use the default
+     *                <code>http://docs.oasis-open.org/ws-dd/ns/discovery/2009/01/rfc3986</code>.
+     * @return future providing probe matches response
+     */
+    ListenableFuture<ProbeMatchesType> sendDirectedProbe(RequestResponseClient rrClient,
+                                                         List<QName> types,
+                                                         List<String> scopes,
+                                                         @Nullable MatchBy matchBy);
+    
     /**
      * Sends a probe message using the given parameters.
      * <p>
@@ -30,13 +46,19 @@ public interface WsDiscoveryClient extends Interceptor {
      * @param probeId of the probe
      * @param types   to probe for
      * @param scopes  to probe for
+     * @param matchBy the scopes matching rule, or null to use the default
+     *                <code>http://docs.oasis-open.org/ws-dd/ns/discovery/2009/01/rfc3986</code>.
+     *
      * @return Future containing the number of matches within the given timeout.
      * @throws MarshallingException on marshalling issues in the outgoing message
      * @throws TransportException   on transport issues in the outgoing message
      * @throws InterceptorException on preprocessing issues with the outgoing message
      * @see #registerHelloByeAndProbeMatchesObserver(HelloByeAndProbeMatchesObserver)
      */
-    ListenableFuture<Integer> sendProbe(String probeId, Collection<QName> types, Collection<String> scopes)
+    ListenableFuture<Integer> sendProbe(String probeId,
+                                        Collection<QName> types,
+                                        Collection<String> scopes,
+                                        @Nullable MatchBy matchBy)
             throws MarshallingException, TransportException, InterceptorException;
 
     /**
@@ -47,6 +69,8 @@ public interface WsDiscoveryClient extends Interceptor {
      * @param probeId    of the probe
      * @param types      to probe for
      * @param scopes     to probe for
+     * @param matchBy the scopes matching rule, or null to use the default
+     *                <code>http://docs.oasis-open.org/ws-dd/ns/discovery/2009/01/rfc3986</code>.
      * @param maxResults number of results to wait for at most
      * @return Future containing the number of matches within the given timeout.
      * @throws MarshallingException on marshalling issues in the outgoing message
@@ -54,22 +78,12 @@ public interface WsDiscoveryClient extends Interceptor {
      * @throws InterceptorException on preprocessing issues with the outgoing message
      * @see #registerHelloByeAndProbeMatchesObserver(HelloByeAndProbeMatchesObserver)
      */
-    ListenableFuture<Integer> sendProbe(String probeId, Collection<QName> types,
-                                        Collection<String> scopes, Integer maxResults)
+    ListenableFuture<Integer> sendProbe(String probeId,
+                                        Collection<QName> types,
+                                        Collection<String> scopes,
+                                        @Nullable MatchBy matchBy,
+                                        Integer maxResults)
             throws MarshallingException, TransportException, InterceptorException;
-
-    /**
-     * Sends a directed probe to a device.
-     *
-     * @param rrClient to send request on
-     * @param types    to probe for
-     * @param scopes   to probe for
-     * @return future providing probe matches response
-     * <p>
-     * TODO LDe: This is inconsistent with the rest, why does it not throw any exceptions?
-     */
-    ListenableFuture<ProbeMatchesType> sendDirectedProbe(RequestResponseClient rrClient, List<QName> types,
-                                                         List<String> scopes);
 
     /**
      * Sends a resolve message for a given endpoint reference.