+++ CHANGELOG.md
@@ -35,10 +35,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reduced log level of network interface information on DPWS framework startup in `org.somda.sdc.dpws.DpwsFrameworkImpl`. (#245)
 - Avoid use of reflection in `org.somda.sdc.biceps.common.MdibTypeValidator` to improve performance. (#248)
 - `org.somda.sdc.dpws.soap.NotificationSink` throws `SoapFaultException`. (#252)
+- `org.somda.sdc.dpws.soap.wsdiscovery.WsDiscoveryClient` and `org.somda.sdc.dpws.client.DiscoveryFilter` to support manual configuration of MatchBy. (#263)
 
 ### Removed
 
 - `org.somda.sdc.glue.common.helper.UrlUtf8` `.decode` and `.encode` methods (#247)
+- `setMatchBy` and `getMatchBy` from `org.somda.sdc.dpws.soap.wsdiscovery.WsDiscoveryTargetService`. (#263)
+- Setting MatchBy for Hello and Bye messages. (#263)
 
 ### Fixed